Finish Grand Exchange functionality. Please help test

+++ src/main/java/com/rs/game/player/Inventory.java
@@ -188,6 +188,8 @@ public final class Inventory {
 		for (Item item : list) {
 			if (item == null)
 				continue;
+			if (!containsItem(item))
+				return false;
 			deleteItem(item);
 		}
 		return true;
